annotation_client/requirements.txt
flask
pandas
dash
dash-renderer 
dash-html-components
dash-core-components
dash-bootstrap-components
plotly
svgtrace
cufflinks
requests
pandas
numpy
matplotlib
Pillow
scikit-image
kaleido
